+++ src/app/(route)/(withLayout)/items/[itemId]/_component/ReviewModal.tsx
@@ -1,19 +1,19 @@
 'use client'
 
-import { ChangeEvent, useState } from 'react'
-import { ReviewInfo, ReviewState } from '@/app/_types/review.type'
-import useAddReview from '@/app/_hook/api/reviews/useAddReview'
-import useEditReview from '@/app/_hook/api/reviews/useEditReview'
 import Modal from '@/app/_components/modal'
+import useAddReview from '@/app/_hook/api/reviews/mutations/useAddReview'
+import useEditReview from '@/app/_hook/api/reviews/mutations/useEditReview'
+import { ReviewInfo, ReviewState } from '@/app/_types/review.type'
 import { cn } from '@/app/_utils/twMerge'
+import { ChangeEvent, useState } from 'react'
 import {
   validateForm,
   validateImage,
   validateImageSize,
 } from '../_utils/validation'
+import ReviewModalForm from './ReviewModalForm'
 import ReviewModalHeader from './ReviewModalHeader'
 import ReviewModalItemDisplay from './ReviewModalItemDisplay'
-import ReviewModalForm from './ReviewModalForm'
 
 interface PropsType {
   setShowReviewModal: React.Dispatch<React.SetStateAction<boolean>>
