+++ src/app/(route)/(withLayout)/mypage/_components/VoteContentTabs.tsx
@@ -1,10 +1,10 @@
 'use client'
 
-import { useUserContentListData } from '@/app/_hook/api/mypage/useContentList'
-import { cn } from '@/app/_utils/twMerge'
-import useGetSearchParam from '@/app/_hook/common/useGetSearchParams'
-import { useClientCookies } from '@/app/_hook/common/useClientCookies'
 import Loading from '@/app/_components/loading'
+import { useUserContentListData } from '@/app/_hook/api/mypage/queries/useContentList'
+import { useClientCookies } from '@/app/_hook/common/useClientCookies'
+import useGetSearchParam from '@/app/_hook/common/useGetSearchParams'
+import { cn } from '@/app/_utils/twMerge'
 import { useEffect } from 'react'
 import { useInView } from 'react-intersection-observer'
 import VoteItem from '../../votes/_component/VoteItem'
