+++ src/app/(route)/(withLayout)/votes/add-vote/_component/VoteItemList.tsx
@@ -1,12 +1,12 @@
 'use client'
 
-import { fetchFavoriteList } from '@/app/_hook/api/votes/useFavoritesList'
-import { useEffect, useState } from 'react'
+import { fetchFavoriteList } from '@/app/_hook/api/votes/queries/useFavoritesList'
 import {
   CurrentFavoriteItemMetadata,
   SaveItemType,
 } from '@/app/_types/saveItem.type'
 import { cn } from '@/app/_utils/twMerge'
+import { useEffect, useState } from 'react'
 import FolderList from './FolderList'
 import VoteItemsSelector from './VoteItemsSelector'
