+++ src/app/_hook/api/votes/mutations/useDeleteVote.ts
@@ -1,8 +1,8 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query'
 import renderToast from '@/app/_utils/toast'
+import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { getCookie } from 'cookies-next'
-import { voteKeys } from '.'
-import { useHandleApiError } from '../../common/useHandleApiError'
+import { voteKeys } from '..'
+import { useHandleApiError } from '../../../common/useHandleApiError'
 
 async function deleteVote(voteId: number): Promise<number> {
   const accessToken = getCookie('accessToken')
